fix more compiler warning

+++ include/Cards/AuraHelper.h
@@ -97,7 +97,7 @@ namespace Cards
 			auto & minions = context.manipulate_.Board().Player(card.GetPlayerIdentifier()).minions_;
 			assert(minions.Get(zone_pos) == context.card_ref_);
 			if (zone_pos > 0) context.new_targets.insert(minions.Get(zone_pos - 1));
-			if (zone_pos < (minions.Size() - 1)) context.new_targets.insert(minions.Get(zone_pos + 1));
+			if (zone_pos < ((int)minions.Size() - 1)) context.new_targets.insert(minions.Get(zone_pos + 1));
 		}
 	};
 	template <typename HandleClass, typename EnchantmentType>
